feat: implement PluginActionSaga to handle plugin action execution
### Changes Made
- Introduced `PluginActionSaga.ts` to manage the execution of plugin actions, including handling blob URLs, action timeouts, and error responses.
- Removed the deprecated `baseExectutePluginSaga.ts`, consolidating functionality into the new saga for improved maintainability.
- Updated related sagas to utilize the new structure, ensuring seamless integration with existing action execution flows.
- Enhanced error handling and logging for better debugging and user feedback during action execution.
- Implemented soft refresh and unload action handling to improve user experience during page transitions.
